+++ include/global.h
@@ -1,0 +1,6 @@
+#ifndef _GLOBAL_H
+#define _GLOBAL_H
+
+#define MAX_IN 80
+
+#endif
